style(split): profile
