+++ backend/matcha/routes/profile.py
@@ -5,7 +5,7 @@ from flask_jwt_extended import (
     get_jwt_identity,
 )
 
-from matcha.services.users import (
+from matcha.services.profile import (
     services_profile,
 )
